fix(lsp): выставлять workspace-контекст в didClose вместо фолбэка в аспекте
didClose приходит с потока диспетчеризации LSP4J без установленного
WorkspaceContextHolder — единственная входная точка без обёртки. Ставим
контекст явно, как в didOpen и остальных обработчиках, вместо
восполнения контекста на уровне EventPublisherAspect: изменения аспекта,
WorkspaceContextHolder и ServerContextProvider откачены, регрессионный
тест переведён на закрытие через BSLTextDocumentService.didClose.
